refactor: Move classes related to COEFFICIENT to a separate file.
